start rcmd

+++ ecom-accounts/src/main/sql/AccounInfo.sql
@@ -1,5 +1,5 @@
 CREATE TABLE ACCOUNT_INFOS(
-            account_id INT PRIMARY KEY NOT NULL ,
+            account_id LONG PRIMARY KEY NOT NULL ,
             user_name VARCHAR(255) ,
             user_pwd INT,
             user_description TEXT,
